L1+L2: Refactor/Major: namespace order; names; driver
Namespaces are re-ordered to be <concept>::<language> instead of
<language>::<concept> to better reflect the multi-language pipeline
nature of the whole compiler.

'parse_tree' is universally renamed to 'ast'. The L1 codegen namespace
hierarchy is flattened slightly, removing its references to an 'ast'
subnamespace that is not necessary to have explicit. (It depends on the
AST no matter how you look at it. To generalize it beyond the existing
AST would be no small matter; besides, the explicit namespace does not
imply a specific AST or make generalizing easier in any way.)

The L2 'main' was getting unwieldy. Now, the main function is very
small, and simply invokes the L2 'driver', which namespace contains
the 'Options' type for invokation options, the argv parser for the
Options structure, and the 'execute' command that handles the execution
of an input given an Options structure.